% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSequenceInputs.R
\name{dartSeqNewInputs}
\alias{dartSeqNewInputs}
\title{dartSeqNewInputs}
\usage{
dartSeqNewInputs(simName, sequenceFileXML, propertyArgs, DARTprogDir,
  DARTprocess, newSequenceFileXML = sequenceFileXML, maxTime = 120,
  userDescBool = FALSE, userDesc = NULL)
}
\arguments{
\item{simName}{Name of your simulation.}

\item{sequenceFileXML}{Name of the xml file containing property(s) of which the arguments you would like to change.}

\item{propertyArgs}{Required inputs into DART simulation. This should be a list with the propertyName as the name and a vector of values as
the argument. (e.g:  list("Maket.Scene.CellDimensions.x" = c(1, 2, 3, 4, 5, 6), "Maket.Scene.CellDimensions.z" = c(2, 4, 6)))}

\item{DARTprogDir}{Path to the 'DART' directory e.g 'C:/User/<username>/DART'.}

\item{DARTprocess}{The DART process(s) you would like to run e.g. directions, maket, phase}

\item{newSequenceFileXML}{Name of the xml file with new propertyArgs defined. Default is to overwrite sequenceFileXML}

\item{maxTime}{Maximum amount of time dart-sequence.bat is allowed to run.}

\item{userDescBool}{If True the user must define a new name in which to store the sequence files. Otherwise the datetime in the
form YYYY_DOY_H.M.S is used.}

\item{userDesc}{If userDescBool is TRUE then this will be a string containing the directory name in which the sequence files will be stored.}
}
\value{
A sequence of files that together can be used input all propertyArgs into a DART simulation.
}
\description{
Create a DART sequence based on the arguents defined.
}
