% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editSequence.R
\name{editSequence}
\alias{editSequence}
\title{editSequence}
\usage{
editSequence(simName, sequenceFileXML, groupNames, propertyNames,
  propertyArgs, DARTprogDir, newSequenceFileXML)
}
\arguments{
\item{simName}{Name of your simulation.}

\item{sequenceFileXML}{Name of the xml file containing property(s) of which the arguments you would like to change.}

\item{groupNames}{Name(s) of the group in which the property whos args you would like to define is}

\item{propertyNames}{Name(s) of the property which you would like to define the DART input for.}

\item{propertyArgs}{Required property input into DART simulation. This should be a vector if changing one arg or a list of vectors
if changing multiple.}

\item{DARTprogDir}{Path to the 'DART' directory e.g 'C:/User/<username>/DART'.}

\item{newSequenceFileXML}{Name of the xml file with new propertyArgs defined. Default is to overwrite sequenceFileXML}
}
\value{
A XML file newSequenceFileXML with the desired arguments.
}
\description{
Change the arguments of an DART sequence XML file.
}
