% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSequenceJobScripts.R
\name{makeSequenceJobScripts}
\alias{makeSequenceJobScripts}
\title{makeSequenceJobScripts}
\usage{
makeSequenceJobScripts(simName, sequenceFileXML, DARTprogDir, DARTprocess,
  maxTime = 120, userDescBool = FALSE, userDesc = NULL)
}
\arguments{
\item{simName}{Name of your simulation.}

\item{sequenceFileXML}{Name of the xml file containing property(s) of which the arguments you want to sequence and input into
DART simulation.}

\item{DARTprogDir}{Path to the 'DART' directory e.g 'C:/User/<username>/DART'.}

\item{DARTprocess}{The DART process(s) you would like to run e.g. directions, maket, phase
Maximum amount of time dart-sequence.bat is allowed to run.}

\item{userDescBool}{If True the user must define a new name in which to store the sequence files. Otherwise the datetime in the
form YYYY_DOY_H.M.S is used.}

\item{userDesc}{If userDescBool is TRUE then this will be a string containing the directory name in which the sequence files will be stored.}
}
\value{
Sequence of folders with required args and a job script that will run the sequenced inputs using a designated DART
process(es).
}
\description{
Create scripts that run DART simulations.
}
