% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSequenceInputs.R
\name{dartSeqNewInputs}
\alias{dartSeqNewInputs}
\title{Select the arguments you would like to run in DART and create a sequence of the arguments.}
\usage{
dartSeqNewInputs(inputPath, groupNames, propertyNames, newArgs, dartDir,
  newPath = inputPath)
}
\arguments{
\item{newPath}{}
}
\description{
Select the arguments you would like to run in DART and create a sequence of the arguments.
}
