% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSequenceInputs.R
\name{dartSeqNewInputs}
\alias{dartSeqNewInputs}
\title{dartSeqNewInputs}
\usage{
dartSeqNewInputs(simName, sequenceFileXML, groupNames, propertyNames,
  propertyArgs, DARTprogDir, DARTprocess,
  newSequenceFileXML = sequenceFileXML, maxTime = 120,
  userDescBool = FALSE, userDesc = NULL)
}
\arguments{
\item{simName}{Name of your simulation.}

\item{sequenceFileXML}{Name of the xml file containing property(s) of which the arguments you would like to change.}

\item{groupNames}{Name(s) of the group in which the property whos args you would like to define is}

\item{propertyNames}{Name(s) of the property which you would like to define the DART input for.}

\item{propertyArgs}{Required property input into DART simulation. This should be a vector if changing one arg or a list of vectors
if changing multiple.}

\item{DARTprogDir}{Path to the 'DART' directory e.g 'C:/User/<username>/DART'.}

\item{DARTprocess}{The DART process(s) you would like to run e.g. directions, maket, phase}

\item{newSequenceFileXML}{Name of the xml file with new propertyArgs defined. Default is to overwrite sequenceFileXML}

\item{maxTime}{Maximum amount of time dart-sequence.bat is allowed to run.}

\item{userDescBool}{If True the user must define a new name in which to store the sequence files. Otherwise the datetime in the
form YYYY_DOY_H.M.S is used.}

\item{userDesc}{If userDescBool is TRUE then this will be a string containing the directory name in which the sequence files will be stored.}
}
\value{
A sequence of files that together can be used input all propertyArgs into a DART simulation.
}
\description{
Create a DART sequence based on the arguents defined.
}
