% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSequenceInputs.R
\name{dartSeqNewInputs}
\alias{dartSeqNewInputs}
\title{dartSeqNewInputs}
\usage{
dartSeqNewInputs(simName, sequenceFileXML, groupNames, propertyNames,
  propertyArgs, DARTprogDir, newSequenceFileXML = sequenceFileXML,
  maxTime = 120)
}
\arguments{
\item{simName}{Name of your simulation.}

\item{sequenceFileXML}{Name of the xml file containing property(s) of which the arguments you would like to change.}

\item{groupNames}{Name(s) of the group in which the property whos args you would like to define is}

\item{propertyNames}{Name(s) of the property which you would like to define the DART input for.}

\item{propertyArgs}{Required property input into DART simulation. This should be a vector if changing one arg or a list of vectors
if changing multiple.}

\item{DARTprogDir}{Path to the 'DART' directory e.g 'C:/User/<username>/DART'.}

\item{newSequenceFileXML}{Name of the xml file with new propertyArgs defined. Default is to overwrite sequenceFileXML}

\item{maxTime}{Maximum amount of time dart-sequence.bat is allowed to run.}
}
\value{
A sequence of files that together can be used input all propertyArgs into a DART simulation.
}
\description{
Create a DART sequence based on the arguents defined.
}
