% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editSequence.R
\name{editSequence}
\alias{editSequence}
\title{editSequence}
\usage{
editSequence(simName, sequenceFileXML, propertyArgs, DARTprogDir,
  newSequenceFileXML)
}
\arguments{
\item{simName}{Name of your simulation.}

\item{sequenceFileXML}{Name of the xml file containing property(s) of which the arguments you would like to change.}

\item{propertyArgs}{Required inputs into DART simulation. This should be a list with the propertyName as the name and a vector of values as
the argument. (e.g:  list("Maket.Scene.CellDimensions.x" = c(1, 2, 3, 4, 5, 6), "Maket.Scene.CellDimensions.z" = c(2, 4, 6)))}

\item{DARTprogDir}{Path to the 'DART' directory e.g 'C:/User/<username>/DART'.}

\item{newSequenceFileXML}{Name of the xml file with new propertyArgs defined. Default is to overwrite sequenceFileXML}
}
\value{
A XML file newSequenceFileXML with the desired arguments.
}
\description{
Change the arguments of an DART sequence XML file.
}
